[UPD] Change timer schema fields from uint8 to int and update API request URL format

+++ backend/schemas/timer-schemas.go
@@ -13,8 +13,8 @@ const (
 )
 
 type TimerActionSpecificHour struct {
-	Hour   uint8 `json:"hour"`
-	Minute uint8 `json:"minute"`
+	Hour   int `json:"hour"`
+	Minute int `json:"minute"`
 }
 
 type TimerReactionGiveTime struct{}
@@ -23,9 +23,9 @@ type TimeAPISTRUCT struct {
 	Year         int    `json:"year"`
 	Month        int    `json:"month"`
 	Day          int    `json:"day"`
-	Hour         uint8  `json:"hour"`
-	Minute       uint8  `json:"minute"`
-	Seconds      uint8  `json:"seconds"`
+	Hour         int    `json:"hour"`
+	Minute       int    `json:"minute"`
+	Seconds      int    `json:"seconds"`
 	MilliSeconds int    `json:"milliSeconds"`
 	DateTime     string `json:"dateTime"`
 	Date         string `json:"date"`
